Replace Either[JsonRpcErrorResponse, JsValue] with distinct response types
This allows ResponseCompanion to be simplified because it now only has to handle
success responses.

The generic response reader is written such that it's biased towards reading a
success result, but if the "error" field is present it will attempt to read an
error (so validation errors will relate to the way in which the message fails
to be a valid error message if "error" is present, but will otherwise will
relate to how it fails to be a valid success message).

Also simplify the construction of JsErrors in the tests and improve the error
messages for invalid params and correlation IDs.